# api/openapi.yaml
openapi: 3.1.0
info:
  title: Phenk
  version: 0.1.0
  summary: Public inboxes for people and agents
  description: |
    Type any name and you have an inbox. No account, no password, no setup:
    mail sent to `anything@<public domain>` lands in the inbox called
    `anything`, and anyone who knows the name can read it. Messages are kept
    for a rolling window and then deleted.

    That makes Phenk useful to two audiences. People use it to receive a
    one-off verification email without handing out their real address.
    Agents and test suites use it to receive mail during an automated flow:
    pick a name, use it, then long-poll `GET /v1/named/{address}/wait` until
    the message arrives. Every message carries any verification codes and
    links detected in it, so the common case needs no parsing at all.

    There are two kinds of address and the difference matters. A **named**
    address is the public inbox above: shared by everybody who knows the name,
    and permanent. A **random** address is unguessable, owned by the session
    that created it, and destroyed on a deadline. Random addresses are an
    optional feature, off unless the operator enables it; `GET /v1/meta`
    says which features this server offers.

    Named addresses can never hold a grant, a webhook, or an API key, because
    there is no one to grant anything to.
  license:
    name: AGPL-3.0-or-later
servers:
  - url: https://api.phenk.example
    description: Production
...
paths:
...
  /v1/identities/{id}/wait:
    parameters:
      - $ref: '#/components/parameters/IdentityId'
      - $ref: '#/components/parameters/Since'
      - name: timeout
        in: query
        description: Seconds to hold the request open. Capped at 120.
        schema: { type: integer, minimum: 1, maximum: 120, default: 30 }
    get:
      operationId: waitForMessages
      tags: [messages]
      summary: Long-poll for the next message
      description: |
        Returns immediately if anything has already arrived after `since`, and
        otherwise holds the request open until something does or the timeout
        expires.

        The already-arrived check happens before any subscription, which is the
        whole point: the gap between creating an address and first waiting on it
        is exactly where real mail lands.

        Messages are returned once parsed, so their subject, preview and
        `extracted` are filled in. If parsing stalls, a message is returned
        after a short grace period anyway, in state `received`.
      responses:
        '200':
          description: |
            Zero or more messages. An empty list with `timed_out: true` means
            nothing arrived; call again with the returned cursor.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WaitResult' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/named:
    post:
      operationId: openNamedInbox
      tags: [named]
      summary: Open a public inbox by name
      description: |
        Resolves a named inbox, creating it if nobody has used the name before.
        The same name always resolves to the same address.

        **Anyone who guesses the name can read this inbox.** Every interface
        showing one is required to say so.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OpenNamedRequest' }
      responses:
        '200':
          description: The inbox, whether it already existed or was just created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Identity' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503':
          description: No public-pool domain can currently hand out addresses.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/named/{address}/messages:
    parameters:
      - $ref: '#/components/parameters/NamedAddress'
      - $ref: '#/components/parameters/Since'
      - $ref: '#/components/parameters/Limit'
    get:
      operationId: listNamedMessages
      tags: [named]
      summary: List a public inbox's messages
      description: No ownership check, by design. There is no owner.
      responses:
        '200':
          description: Messages in arrival order, oldest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageList' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/named/{address}/stream:
    parameters:
      - $ref: '#/components/parameters/NamedAddress'
      - $ref: '#/components/parameters/Since'
    get:
      operationId: streamNamedInbox
      tags: [named]
      summary: Server-sent event stream for a public inbox
      responses:
        '200':
          description: An event stream.
          content:
            text/event-stream:
              schema: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    IdentityId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    NamedAddress:
      name: address
      in: path
      required: true
      description: A local part, or a full address on a public-pool domain.
      schema: { type: string }
    Since:
      name: since
      in: query
      description: Return only what arrived after this cursor.
      schema: { type: integer, format: int64, minimum: 0, default: 0 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    BadRequest:
      description: The request was malformed or asked for something disallowed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        No such thing, or it is not yours. The two are deliberately
        indistinguishable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: Rate limited.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    OpenNamedRequest:
      type: object
      required: [local_part]
      properties:
        local_part:
          type: string
          description: |
            Lower-cased and stripped of any `+tag` before use. Must match
            `^[a-z0-9][a-z0-9._-]{2,63}$`, must not be a reserved name such as
            `postmaster`, and must not look like a generated address.
      additionalProperties: false
    Identity:
      type: object
      required: [id, address, local_part, domain, kind, state, cursor, public, created_at, quota]
      properties:
        id: { type: string, format: uuid }
        address: { type: string }
        local_part: { type: string }
        domain: { type: string }
        kind: { type: string, enum: [random, named] }
        state:
          type: string
          enum: [active, expiring, expired, purged, reserved]
        cursor:
          type: integer
          format: int64
          description: The sequence number of the newest message. Pass it as `since`.
        public:
          type: boolean
          description: |
            True for named inboxes. Anyone who guesses the name can read them,
            and every interface showing one must say so plainly.
        created_at: { type: string, format: date-time }
        expires_at:
          type: string
          format: date-time
          description: Absent for named inboxes, which are permanent.
        retention_hours:
          type: integer
          description: Rolling per-message retention. Named inboxes only.
        quota: { $ref: '#/components/schemas/Quota' }
    Quota:
      type: object
      required: [messages, bytes, used_messages, used_bytes]
      properties:
        messages: { type: integer }
        bytes: { type: integer, format: int64 }
        used_messages: { type: integer }
        used_bytes: { type: integer, format: int64 }
    MessageList:
      type: object
      required: [messages, cursor]
      properties:
        messages:
          type: array
          items: { $ref: '#/components/schemas/MessageSummary' }
        cursor:
          type: integer
          format: int64
          description: Pass as `since` on the next call.
    WaitResult:
      allOf:
        - $ref: '#/components/schemas/MessageList'
        - type: object
          required: [timed_out]
          properties:
            timed_out:
              type: boolean
              description: True when the wait expired with nothing new.
    MessageSummary:
      type: object
      required: [id, seq, subject, preview, from, received_at, size_bytes, state, attachment_count, auth]
      properties:
        id: { type: string, format: uuid }
        seq: { type: integer, format: int64 }
        subject: { type: string }
        preview: { type: string }
        from: { $ref: '#/components/schemas/Address' }
        received_at: { type: string, format: date-time }
        sent_at:
          type: string
          format: date-time
          description: The Date header, when the message carried a usable one.
        size_bytes: { type: integer, format: int64 }
        state:
          type: string
          enum: [received, parsed, failed]
          description: |
            `received` means the message is stored but not yet parsed, so
            subject and preview may still be empty. `failed` means it could not
            be parsed; the raw message is still downloadable.
        attachment_count: { type: integer }
        auth: { $ref: '#/components/schemas/AuthResults' }
        extracted:
          $ref: '#/components/schemas/Extracted'
    Extracted:
      type: object
      required: [codes, links]
      description: |
        What an automated reader usually came for, found by rules rather than
        a model. Present once a message is parsed. Detection is deliberately
        conservative — a code is only reported when the words around it say it
        is one — so an empty list is common and a reported code is usually
        right. It is still a guess: `context` shows where it came from.
      properties:
        codes:
          type: array
          description: Verification codes, most likely first. At most three.
          items: { $ref: '#/components/schemas/DetectedCode' }
        links:
          type: array
          description: |
            http and https links, account-acting ones first and unsubscribe
            links last. At most ten.
          items: { $ref: '#/components/schemas/DetectedLink' }
    DetectedCode:
      type: object
      required: [value, context]
      properties:
        value:
          type: string
          description: The code with any grouping removed, as a form expects it.
        context:
          type: string
          description: The line the code was found on.
    DetectedLink:
      type: object
      required: [url, text, kind]
      properties:
        url: { type: string, format: uri }
        text:
          type: string
          description: The link's visible text, when it had any.
        kind:
          type: string
          enum: [verify, unsubscribe, other]
          description: |
            `verify` acts on an account: confirm, verify, activate, sign in,
            reset a password, accept an invitation. `unsubscribe` stops mail.
...
    Address:
      type: object
      required: [address]
      properties:
        name: { type: string }
        address: { type: string }
...
    AuthResults:
      type: object
      required: [spf, dkim, dmarc]
      description: |
        `none` means the check was not evaluated or the sender published
        nothing to check against. It is not a pass.
      properties:
        spf: { $ref: '#/components/schemas/AuthResult' }
        dkim: { $ref: '#/components/schemas/AuthResult' }
        dmarc: { $ref: '#/components/schemas/AuthResult' }
    AuthResult:
      type: string
      enum: [none, pass, fail, softfail, neutral, temperror, permerror]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }